pyproject.toml
[project]
name = "test-whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown>=3.10",
    "ollama>=0.6.1",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.1",
    "pytubefix>=10.3.4",
    "rich>=14.2.0",
    "weasyprint>=66.0",
]
[project.scripts]
summarize-video = "summarize_video.cli:main"
